requirements.txt
unsloth
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision
torchaudio
transformers>=4.40.0
pillow
bitsandbytes
unsloth_zoo
triton
peft
accelerate
safetensors